Split storages into separate files
