+++ src/lib.rs
@@ -205,9 +205,9 @@ pub use shred::{Dispatcher, DispatcherBuilder, Fetch, FetchId, FetchIdMut,
 #[cfg(not(target_os = "emscripten"))]
 pub use shred::{AsyncDispatcher};
 
-pub use storage::{BTreeStorage, CheckStorage, DenseVecStorage, DistinctStorage, FlaggedStorage,
-                  HashMapStorage, InsertResult, MaskedStorage, NullStorage, ReadStorage, Storage,
-                  UnprotectedStorage, VecStorage, WriteStorage};
+pub use storage::{BTreeStorage, CheckStorage, DenseVecStorage, DistinctStorage, Entry, FlaggedStorage,
+                  HashMapStorage, InsertResult, MaskedStorage, NullStorage, ReadStorage, RestrictedStorage,
+                  Storage, UnprotectedStorage, VecStorage, WriteStorage};
 pub use world::{Component, CreateIter, CreateIterAtomic, EntitiesRes, Entity, EntityBuilder,
                 Generation, LazyUpdate, World};
