+++ src/storage/mod.rs
@@ -1,6 +1,7 @@
 //! Storage types
 
-pub use self::check::{RestrictedStorage, CheckStorage, Entry};
+pub use self::check::CheckStorage;
+pub use self::restrict::{RestrictedStorage, Entry};
 pub use self::data::{ReadStorage, WriteStorage};
 #[cfg(feature = "serialize")]
 pub use self::ser::{MergeError, PackedData};
@@ -20,6 +21,7 @@ use shred::Fetch;
 use {Component, EntitiesRes, Entity, Index, Join, ParJoin};
 
 mod check;
+mod restrict;
 mod data;
 #[cfg(feature = "serialize")]
 mod ser;
